requirements.txt
numpy
pyte==0.8.1
openai
python-dotenv